refactor(actions): modularize double-click and right-click from click
