+++ src/actions/index.ts
@@ -2,10 +2,12 @@ export * from './blur';
 export * from './click';
 export * from './cookie';
 export * from './debug';
+export * from './double-click';
 export * from './go';
 export * from './log';
 export * from './pause';
 export * from './reload';
+export * from './right-click';
 export * from './screenshot';
 export * from './scroll';
 export * from './select';
